pyproject.toml
[tool.poetry]
name = "self-healing-code"
version = "0.0.0"
description = "Agent-based, self-healing coding platform"
authors = ["You <you@example.com>"]
packages = [{ include = "apps" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112"
uvicorn = {extras = ["standard"], version = "^0.30"}
asyncpg = "^0.30"
sqlmodel = "^0.0.16"
grpclib = "^0.4.7"
pydantic = "^2.7"
alembic = "^1.13"
prometheus-client = "^0.20"
pygit2 = "^1.14"
protobuf = ">=5.26.1,<6"
grpcio = "1.65.5"
grpcio-tools = "1.65.5"
aiofiles = "^23.2"
aiohttp = "^3.9"
qdrant-client = "^1.9"
transformers = "^4.44"
sentence-transformers = "^3.0"
scikit-learn = "^1.5"
tree-sitter = "^0.22"
sqlite-utils = "^3.36"
httpx = "^0.27"
tenacity = "^8.3"
confluent-kafka = "^2.4"
transitions = "^0.9"
python-dotenv = "^1.0"
openai = "^1.35"
backoff = "^2.2"
tiktoken = "^0.6"
anyio = "^4.4"
orjson = "^3.10"
networkx = "^3.3"
radon = "^6.0"
unidiff = "^0.7"
GitPython = "^3.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7"
pytest-asyncio = "^0.23"
pytest = "^8.2"

[tool.poetry.group.ci.dependencies]
pytest = "^8.2"
coverage = {extras = ["toml"], version = "^7.5"}
pytest-testmon = "^2.1"
ruff = "^0.4"
black = "^24.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"